feat(abstract-utxo): move replay protection and signing to transaction/fixedScript
Move files to their appropriate locations:
- Move replayProtection.ts to transaction/fixedScript/
- Move sign.ts to transaction/fixedScript/
- Update all import references throughout the codebase

Issue: BTC-2806
