+++ modules/abstract-utxo/src/index.ts
@@ -2,8 +2,8 @@ export * from './abstractUtxoCoin';
 export * from './address';
 export * from './config';
 export * from './recovery';
-export * from './replayProtection';
-export * from './sign';
+export * from './transaction/fixedScript/replayProtection';
+export * from './transaction/fixedScript/sign';
 
 export { UtxoWallet } from './wallet';
 export * as descriptor from './descriptor';
