+++ modules/abstract-utxo/src/transaction/fixedScript/index.ts
@@ -4,3 +4,5 @@ export { parseTransaction } from './parseTransaction';
 export { CustomChangeOptions } from './parseOutput';
 export { verifyTransaction } from './verifyTransaction';
 export { signTransaction } from './signTransaction';
+export * from './sign';
+export * from './replayProtection';
