+++ modules/abstract-utxo/src/transaction/fixedScript/sign.ts
